Handle session cookies (logins while logged in) better
If we have a session cookie already, then the NIAuth server does not
give us the 403 with the info for the SRP handshake, because from its
point of view we're already logged in. Handle that case now. Also, set
'same-origin' credentials in the fetch() calls so that we will start
using session cookies.

Also add the ability to refresh the client's "logged in" state from the
session cookie (e.g., on page refresh).

Add marginal handling for the "logging in as a different user" case.
Right now this gets pushed onto the caller. Ideally we would log out and
log back in behind the scenes, but the login function needs to be
refactored to act as a proper state machine before that's not daunting
to add.

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>